FirewallEngine: add module encapsulating WFP operations
The module initializes session to WFP engine, sets up a sublayer
where filters will be added and exposes simple API to add
and remove filters. Once filter is added to the module,
it's added to the engine as well.

This API is subject to change, but it will be a base to iterate
over as new features are added.